+++ tempest_fastapi_sdk/__init__.py
@@ -151,7 +151,7 @@ from tempest_fastapi_sdk.webpush import (
     WebPushSubscriptionSchema,
 )
 
-__version__: str = "0.15.0"
+__version__: str = "0.16.0"
 
 __all__: list[str] = [
     "CEP",